BE.Copyright.Manager/requirements.txt
python-dotenv
flask
Flask-Cors
blueprint
numpy==1.22.1
sentence_transformers
sklearn
mysql-connector-python
pymongo
opencv-python
matplotlib
Pillow
transformers
pymysql